refactor(ingest): relocate pipeline logic and schemas
